requirements.txt
motor
python-dotenv
httpx
pydantic
protobuf
websockets
upstox-python-sdk
redis
